Move performance fakes into browser package
This is for a few reasons:

- We need the stricter performance stuff to be defined in the browser
  package and used by the test utilities package
  That implies a dependency on the browser package in test utilities,
  which is clearly not right
- The performance fake stuff is only relevant to the browser now; it may
  be relevant to other future platforms (looking at you electron) but
  that can be solved by making a new shared package, which could export
  testing stuff only when running the unit tests
  There's no need to create that shared package now, so it living in
  the browser package makes sense at the moment
- It's easier to make changes when these interfaces live in the same
  place as the tests — Jest seems to cache TS compilation across package
  boundaries more than it does between local files, so updating the
  interface in test utilities isn't always reflected until you restart
  Jest (if you're running it in watch mode)
  This is probably a bug in Jest's TS compilation caching and it's
  definitely only a minor annoyance, but it _seems_ more reliable when
  the files are co-located with the tests that use them
  Given how coupled these fake interfaces are with the browser package
  (i.e. we use 'performance' all over the place), it's quite likely that
  they will be modified more than other test utilities — something in
  core changing implies updating all platforms as well, so is a bigger
  deal than a browser-specific change

Yes, I did just move these around in a previous branch; no we aren't
talking about that :^)
