+++ packages/platforms/browser/tests/utilities/performance-entry-list.ts
