+++ packages/platforms/browser/tests/utilities/performance-entry.ts
