+++ packages/platforms/browser/tests/utilities/performance-observer-entry-list.ts
