+++ packages/platforms/browser/tests/utilities/performance-observer-manager.ts
