+++ packages/platforms/browser/tests/utilities/performance.ts
