+++ packages/test-utilities/lib/index.ts
@@ -7,7 +7,6 @@ export { default as createTestClient } from './create-test-client'
 export { default as IncrementingClock } from './incrementing-clock'
 export { default as InMemoryDelivery } from './in-memory-delivery'
 export { default as InMemoryProcessor } from './in-memory-processor'
-export * from './performance-observer'
 export { default as resourceAttributesSource } from './resource-attributes-source'
 export { default as spanAttributesSource } from './span-attributes-source'
 export { default as StableIdGenerator } from './stable-id-generator'
